Move TermVerson action outside of admin/cluster namespace
